test refactor: test opportunity program post
